+++ apps/hashdotdev/src/components/page-layout.tsx
@@ -2,10 +2,10 @@ import { Box } from "@mui/material";
 import { useTheme } from "@mui/system";
 import { FunctionComponent, ReactNode } from "react";
 
+import { Footer } from "./footer";
 import { HiringBanner } from "./hiring-banner";
+import { Navbar } from "./navbar";
 import { PreFooter } from "./pre-footer";
-import { Footer } from "./xxx-footer";
-import { Navbar } from "./xxx-navbar";
 
 // @todo extract NavLink component
