+++ apps/hashdotdev/src/pages/blog/index.page.tsx
@@ -14,9 +14,9 @@ import { ComponentProps, Fragment, FunctionComponent } from "react";
 import { BlogPostAuthor, BlogPostPagePhoto } from "../../components/blog-post";
 import { GradientContainer } from "../../components/gradient-container";
 import { FaIcon } from "../../components/icons/fa-icon";
+import { Link } from "../../components/link";
 import { PageLayout } from "../../components/page-layout";
 import { Subscribe } from "../../components/pre-footer";
-import { Link } from "../../components/xxx-link";
 import { parseNameFromFileName } from "../../util/client-mdx-util";
 import { getAllPages, Page } from "../../util/mdx-util";
 import { NextPageWithLayout } from "../../util/next-types";
